move partitioning to separate class
