+++ src/main/scala/com/appsflyer/spark/bigquery/streaming/BigQueryStreamWriter.scala
@@ -1,6 +1,7 @@
 package com.appsflyer.spark.bigquery.streaming
 
 import com.appsflyer.spark.bigquery.BigQueryServiceFactory
+import com.appsflyer.spark.utils.BigQueryPartitionUtils
 import com.google.api.client.googleapis.json.GoogleJsonResponseException
 import com.google.api.services.bigquery.Bigquery
 import com.google.api.services.bigquery.model.{Table, TableDataInsertAllRequest, TableReference, TimePartitioning}
@@ -54,26 +55,8 @@ class BigQueryStreamWriter(fullyQualifiedOutputTableId: String, batchSize: Int,
   }
 
   override def process(value: String): Unit = {
-
     if(isPartitionedByDay) {
-      val datasetId = targetTable.getDatasetId
-      val projectId: String = targetTable.getProjectId
-      val tableName = targetTable.getTableId
-      try {
-        logger.info("Creating Time Partitioned Table")
-        val table = new Table()
-        table.setTableReference(targetTable)
-        val timePartitioning = new TimePartitioning()
-        timePartitioning.setType("DAY")
-        timePartitioning.setExpirationMs(DEFAULT_TABLE_EXPIRATION_MS)
-        table.setTimePartitioning(timePartitioning)
-        val request = bqService.tables().insert(projectId, datasetId, table)
-        val response = request.execute()
-      } catch {
-        case e: GoogleJsonResponseException if e.getStatusCode == 409 =>
-          logger.info(s"$projectId:$datasetId.$tableName already exists")
-        case NonFatal(e) => throw e
-      }
+      BigQueryPartitionUtils.createBigQueryPartitionedTable(targetTable,bqService)
     }
     if (rowIndex < batchSize) {
       rows(rowIndex).setJson(gson.fromJson(value, targetClass)).setInsertId(s"${batchId}_${rowIndex}")
